style: Change building notification comment
We get to the point of building the notification only when the user
is OK with it. We would not even begin collecting notification info
if the user was not OK with it, let alone building the notification.
So it is inaccurate to say that we will build the notification
if the user is OK with it.